refactor(client-core): generalize the connection controller over its client
