+++ packages/client/core/src/index.ts
@@ -3,6 +3,7 @@
  * (docs/ARCHITECTURE.md#packages--repo-layout); the event stream flows through its subscription hooks.
  */
 export * from './client';
+export * from './connection-controller';
 export * from './conversation';
 export * from './conversation-store';
 export * from './react';
